readme-install: Add a note for non GNU and BusyBox only systems
Such systems have no -r option when symlink a file, but it's needed
to work with Arch ABS. That could be fixed in several ways but because
I'm lazy and unsure which solution may best I keep it as it is until
some Guru give a hint or the complaints are still growing.

The options I had considered are:
- Remove the -r option from the Makefile and don't use in the PKGBUILD
  the Makefile but own commands
- Add an extra target "pkg" to the Makefile which has the -r option and
  remove it from the normal "install" target

In any case is it somehow duplicated stuff, so is there a better
solution?