Oops! Fixes suggested/found by running shellcheck
Due to the lack of nice auto tests is the result difficult to rate,
apparently it works. But it is more than poorly tested.

And Wayland is completely untested!

Would be nice to have it in the next release, when ever it happens.
But I would be very reluctant to do that without a couple of positive
feedback that nothing is broken. Escpecially the Wayland fix.

Patches, comments and reports are welcome!